Create fake pdfs via pdf-lib and upload to bucket
